pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linalg-viz"
version = "0.1.0"
description = "A linear algebra visualization library for learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "linalg-viz contributors"}
]
keywords = ["linear algebra", "visualization", "education", "math"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "numpy>=1.21.0",
    "pygame>=2.1.0",
    "PyOpenGL>=3.1.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
labels = [
    "freetype-py>=2.3.0",
]

[project.urls]
Homepage = "https://github.com/Jeditrix/linalg-viz"
Documentation = "https://github.com/Jeditrix/linalg-viz#readme"
Repository = "https://github.com/Jeditrix/linalg-viz"

[tool.setuptools.packages.find]
where = ["."]
include = ["linalg_viz*"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
